+++ src/pynetim/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__version__ = "0.5.0"
+__version__ = "0.5.3"
 __author__ = "Zhang Kaijing"
 
 from . import graph